feat: making it our own

+++ lib/template.h
@@ -8,7 +8,7 @@ using namespace std;
 
 template <class T>
 using vec = vector<T>;
-using ll = int64_t;
+using ll = long long;
 using pii = array<int, 2>;
 using vi = vec<int>;
